Sam and substrings
Samantha and Sam are playing a numbers game. Given a number as a string, no leading zeros, determine the sum of all integer values of substrings of the string. For example, if the string is 42, the substrings are 4,2  and 42. Their sum is 48.

Given an integer as a string, sum all of its substrings cast as integers. As the number may become large, return the value modulo  10^9+7.